+++ .vscode/settings.json
@@ -6,10 +6,12 @@
         "champlist",
         "creamycogs",
         "eune",
+        "Ezreal",
         "gameid",
         "leaguecog",
         "leagueset",
         "Malarne",
+        "mixinmeta",
         "njhargis",
         "puuid",
         "redbot",
